+++ src/features/ProductList/hooks/useProductList.ts
@@ -1,7 +1,6 @@
-import { useEffect, useState, useCallback } from 'react';
+import { useEffect, useState } from 'react';
 
-import { getProductList } from '@/api/product';
-import { useApiRequest } from '@/shared/hooks/useApiRequest';
+import { useProductListRequest } from './useProductListRequest';
 
 import { Product } from '../types/Product';
 
@@ -9,7 +8,11 @@ export const useProductList = () => {
   const [product, setProduct] = useState<Product[]>([]);
   const [categorySelect, setCategorySelect] = useState('전체');
   const [priceSelect, setPriceSelect] = useState('전체');
-  const { isLoading, handleRequest } = useApiRequest();
+  const { fetchProductData, isLoading } = useProductListRequest(
+    setProduct,
+    priceSelect,
+    categorySelect
+  );
 
   const handleCategorySelect = (category: string) => {
     setCategorySelect(category);
@@ -18,24 +21,6 @@ export const useProductList = () => {
   const handlePriceSelect = (price: string) => {
     setPriceSelect(price);
   };
-
-  const fetchProductData = useCallback(async () => {
-    return handleRequest(
-      () =>
-        getProductList({
-          page: 0,
-          size: 20,
-          sort: priceSelect !== '전체' && priceSelect ? `price,${priceSelect}` : '',
-          category: categorySelect === '전체' ? '' : categorySelect,
-        }),
-      (data) => {
-        setProduct(data);
-        return data;
-      },
-      [],
-      { delay: 2000 }
-    );
-  }, [categorySelect, handleRequest, priceSelect]);
 
   useEffect(() => {
     fetchProductData();
