+++ src/features/ProductList/hooks/useProductListRequest.ts
@@ -1,0 +1,34 @@
+import { useCallback } from 'react';
+
+import { getProductList } from '@/api/product';
+import { useApiRequest } from '@/shared/hooks/useApiRequest';
+
+import { Product } from '../types/Product';
+
+export const useProductListRequest = (
+  setProduct: (data: Product[]) => void,
+  categorySelect: string,
+  priceSelect: string
+) => {
+  const { handleRequest, isLoading } = useApiRequest();
+
+  const fetchProductData = useCallback(async () => {
+    return handleRequest(
+      () =>
+        getProductList({
+          page: 0,
+          size: 20,
+          sort: priceSelect !== '전체' && priceSelect ? `price,${priceSelect}` : '',
+          category: categorySelect === '전체' ? '' : categorySelect,
+        }),
+      (data) => {
+        setProduct(data);
+        return data;
+      },
+      [],
+      { delay: 2000 }
+    );
+  }, [categorySelect, handleRequest, priceSelect, setProduct]);
+
+  return { fetchProductData, isLoading };
+};
